refactor choice item: split components for answerValueSet and answerOption
